Finish `Binary/PE` (#15)
* moved symbols -> win32/symbols

* bumped cmake to 3.26 to use ASM_MARMASM

* [pwn::binary::pe] added exception + tests

* [pwn::binary::pe] fixed delay load and tests

* [pwn::binary::pe] added relocations + string names

* [ci] restored arm64 

* [binary::pe] implementation of most sections

* [binary::pe] added battery test

+++ Modules/CTF/Source/Win32/Remote.cpp
@@ -82,7 +82,7 @@ CTF::Remote::send_internal(_In_ std::vector<u8> const& out)
         Utils::hexdump(out);
     }
 
-    return Ok(out.size());
+    return Ok((usize)out.size());
 }
